Refactor SelectionService
- Remove unnecessary Types.SelectionService binding and bind SelectionService directly instead
- Move `SelectionService` into base package. A lot of feature modules are dependent on this service => they would no longer work if the selectModule is not bound or customized
